move test bases into own project
