+++ 1.README_SolutionStructure.md
@@ -44,7 +44,7 @@ number of projects grows.
 The numbers in the folders allow us to set the folder order to match the layering metaphor instead of being alphabetic. 
 
 The shared folder is for meta-infrastructure that is needed across projects. For example, dependency injection config, data contracts, and possibly logging. 
-Do not be tempted to put app functionality in this folder
+Do not be tempted to put app functionality in this folder.
 
 #Project Name Prefixes
 Adding some namespacing to our projects (e.g. Accessors.Project or Tests.Project) allows our underlying file structure to
