+++ TestingExemplar.sln
@@ -28,6 +28,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "0.Documentation", "0.Docume
 		Shared.DatabaseContext\DTOs\README_DataTransferObjects.md = Shared.DatabaseContext\DTOs\README_DataTransferObjects.md
 		Shared.DependencyInjectionKernel\README_DependencyInjection.md = Shared.DependencyInjectionKernel\README_DependencyInjection.md
 		Shared.DependencyInjectionKernel\README_DependencyInjectionKernel.md = Shared.DependencyInjectionKernel\README_DependencyInjectionKernel.md
+		README_NamingPatterns.md = README_NamingPatterns.md
 		Tests.ManagerTests\README_TestClassReuse.md = Tests.ManagerTests\README_TestClassReuse.md
 		Test.NUnitExtensions\README_TestPrefixes.md = Test.NUnitExtensions\README_TestPrefixes.md
 	EndProjectSection
@@ -41,6 +42,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared.DependencyInjectionKernel", "Shared.DependencyInjectionKernel\Shared.DependencyInjectionKernel.csproj", "{5F633149-4C66-4D34-98AB-28EE4662B442}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.NUnitExtensions", "Test.NUnitExtensions\Test.NUnitExtensions.csproj", "{3AC5C65A-6F68-479C-A078-3D2E573111BB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests.TestBases", "Tests.TestBases\Tests.TestBases.csproj", "{EBD233E2-509F-4A6A-B43E-034986583AE2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -84,6 +87,10 @@ Global
 		{3AC5C65A-6F68-479C-A078-3D2E573111BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3AC5C65A-6F68-479C-A078-3D2E573111BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3AC5C65A-6F68-479C-A078-3D2E573111BB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EBD233E2-509F-4A6A-B43E-034986583AE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EBD233E2-509F-4A6A-B43E-034986583AE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EBD233E2-509F-4A6A-B43E-034986583AE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EBD233E2-509F-4A6A-B43E-034986583AE2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -98,6 +105,7 @@ Global
 		{AF26B020-F939-4353-A41D-F6C0F87D0B81} = {CA030337-D04B-4393-9869-897FA379EC1B}
 		{5F633149-4C66-4D34-98AB-28EE4662B442} = {E1966F97-64AF-4CA1-9CDB-DA841C9491B9}
 		{3AC5C65A-6F68-479C-A078-3D2E573111BB} = {3B08858C-6FD1-47E7-824C-0B3F1D234455}
+		{EBD233E2-509F-4A6A-B43E-034986583AE2} = {3B08858C-6FD1-47E7-824C-0B3F1D234455}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1292F67E-01C7-4886-997A-51E54049301C}
