+++ Tests.AccessorTests/packages.config
@@ -5,5 +5,5 @@
   <package id="JustMock" version="2018.1.117.2" targetFramework="net461" />
   <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net461" />
   <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net461" />
-  <package id="NUnit" version="3.9.0" targetFramework="net461" />
+  <package id="NUnit" version="3.10.1" targetFramework="net461" />
 </packages>
