+++ Tests.ManagerTests/packages.config
@@ -6,6 +6,6 @@
   <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net461" />
   <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net461" />
   <package id="Ninject" version="3.3.4" targetFramework="net461" />
-  <package id="NUnit" version="3.9.0" targetFramework="net461" />
+  <package id="NUnit" version="3.10.1" targetFramework="net461" />
   <package id="NUnit3TestAdapter" version="3.10.0" targetFramework="net461" />
 </packages>
