+++ Tests.TestBases/Tests.TestBases.csproj
@@ -5,11 +5,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{3AC5C65A-6F68-479C-A078-3D2E573111BB}</ProjectGuid>
+    <ProjectGuid>{EBD233E2-509F-4A6A-B43E-034986583AE2}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Test.NUnitExtensions</RootNamespace>
-    <AssemblyName>Test.NUnitExtensions</AssemblyName>
+    <RootNamespace>Tests.TestBases</RootNamespace>
+    <AssemblyName>Tests.TestBases</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <NuGetPackageImportStamp>
@@ -33,11 +33,19 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL">
+      <HintPath>..\packages\EntityFramework.6.2.0\lib\net45\EntityFramework.dll</HintPath>
+    </Reference>
+    <Reference Include="EntityFramework.SqlServer, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL">
+      <HintPath>..\packages\EntityFramework.6.2.0\lib\net45\EntityFramework.SqlServer.dll</HintPath>
+    </Reference>
     <Reference Include="nunit.framework, Version=3.10.1.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
       <HintPath>..\packages\NUnit.3.10.1\lib\net45\nunit.framework.dll</HintPath>
     </Reference>
     <Reference Include="System" />
+    <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Transactions" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -46,12 +54,23 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="TestFixture_Prefixed.cs" />
+    <Compile Include="AccessorTestBase.cs" />
+    <Compile Include="ManagerTestBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="App.config" />
     <None Include="packages.config" />
-    <None Include="README_TestPrefixes.md" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Shared.DatabaseContext\Shared.DatabaseContext.csproj">
+      <Project>{6886BD35-1655-456B-A4C2-6B57BD557254}</Project>
+      <Name>Shared.DatabaseContext</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Tests.DataPrep\Tests.DataPrep.csproj">
+      <Project>{458ED4A8-E0F2-4089-94CB-20B0EFDD4B46}</Project>
+      <Name>Tests.DataPrep</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
