+++ Tests.TestBases/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="EntityFramework" version="6.2.0" targetFramework="net461" />
   <package id="NUnit" version="3.10.1" targetFramework="net461" />
 </packages>
